render/blender_requirements.txt
asciitree==0.3.3
fasteners==0.16.3
fft-conv-pytorch==1.1.3
gin-config==0.4.0
numcodecs==0.9.1
numpy==1.21.4
Pillow==8.3.2
six==1.16.0
tabulate==0.8.9
tqdm==4.62.3
typing_extensions==4.0.1
zarr==2.10.3
